+++ src/test/kotlin/com/zepben/evolve/database/sqlite/cim/upgrade/changesets/network/ChangeSet57NetworkValidator.kt
@@ -8,8 +8,11 @@
 
 package com.zepben.evolve.database.sqlite.cim.upgrade.changesets.network
 
+import com.zepben.evolve.database.getNullableDouble
 import com.zepben.evolve.database.paths.DatabaseType
 import com.zepben.evolve.database.sqlite.cim.upgrade.changesets.ChangeSetValidator
+import org.hamcrest.MatcherAssert.assertThat
+import org.hamcrest.Matchers.equalTo
 import java.sql.Statement
 
 object ChangeSet57NetworkValidator : ChangeSetValidator(DatabaseType.NETWORK_MODEL, 57) {
@@ -19,7 +22,8 @@ object ChangeSet57NetworkValidator : ChangeSetValidator(DatabaseType.NETWORK_MOD
     //
 
     override fun setUpStatements(): List<String> = listOf(
-        "INSERT INTO tap_changer_controls (mrid, name, description, num_diagram_objects, location_mrid, num_controls, discrete, mode, monitored_phase, target_deadband, target_value, enabled, max_allowed_target_value, min_allowed_target_value, rated_current, terminal_mrid, limit_voltage, line_drop_compensation, line_drop_r, line_drop_x, reverse_line_drop_r, reverse_line_drop_x, forward_ldc_blocking, time_delay, co_generation_enabled) VALUES ('tap_changer_control_mrid', 'name', 'description', 1, 'location_mrid', 2, true, 'mode', 'monitored_phase', 3, 4, true, 5, 6, 7, 'terminal_mrid', 8, true, 9, 10, 11, 12, true, 13, true)"
+        "INSERT INTO tap_changer_controls (mrid, name, description, num_diagram_objects, location_mrid, num_controls, discrete, mode, monitored_phase, target_deadband, target_value, enabled, max_allowed_target_value, min_allowed_target_value, rated_current, terminal_mrid, limit_voltage, line_drop_compensation, line_drop_r, line_drop_x, reverse_line_drop_r, reverse_line_drop_x, forward_ldc_blocking, time_delay, co_generation_enabled) VALUES ('tap_changer_control_mrid', 'name', 'description', 1, 'location_mrid', 2, true, 'mode', 'monitored_phase', 3, 4, true, 5, 6, 7, 'terminal_mrid', 8, true, 9, 10, 11, 12, true, 13, true)",
+        "INSERT INTO ac_line_segments (mrid, name, description, num_diagram_objects, location_mrid, num_controls, normally_in_service, in_service, commissioned_date, base_voltage_mrid, per_length_sequence_impedance_mrid) VALUES ('acls1', '', '', 1, 'loc1', 10, true, true, 'comm1', 'bv1', 'plsi_mrid_1');",
     )
 
     override fun populateStatements(): List<String> = listOf(
@@ -29,16 +33,21 @@ object ChangeSet57NetworkValidator : ChangeSetValidator(DatabaseType.NETWORK_MOD
         "INSERT INTO static_var_compensators (mrid, name, description, num_diagram_objects, location_mrid, num_controls, normally_in_service, in_service, commissioned_date, base_voltage_mrid, control_enabled, regulating_control_mrid, capacitive_rating, inductive_rating, q, svc_control_mode, voltage_set_point) VALUES ('mrid', 'name', 'description', 1, 'location_mrid', 2, true, true, 'commissioned_date', 'base_voltage_mrid', true, 'regulating_control_mrid', 3.0, 4.0, 5.0, 'svc_control_mode', 6);",
         "INSERT INTO battery_units_battery_controls (battery_unit_mrid, battery_control_mrid) VALUES ('battery_unit_mrid', 'battery_control_mrid');",
         "INSERT INTO end_devices_end_device_functions (end_device_mrid, end_device_function_mrid) VALUES ('end_device_mrid', 'end_device_function_mrid');",
-    )
+        "INSERT INTO per_length_phase_impedances (mrid, name, description, num_diagram_objects) VALUES ('plpi_mrid', 'name', 'description', 0);",
+        "INSERT INTO phase_impedance_data (per_length_phase_impedance_mrid, from_phase, to_phase, b, g, r, x) VALUES ('plpi_mrid', 'A', 'B', 1.0, 2.0, 3.0, 4.0);",
+        "INSERT INTO ac_line_segments (mrid, name, description, num_diagram_objects, location_mrid, num_controls, normally_in_service, in_service, commissioned_date, base_voltage_mrid, per_length_impedance_mrid) VALUES ('acls2', '', '', 1, 'loc1', 10, true, true, 'comm1', 'bv1', 'plpi_mrid_2');",
+
+        )
 
     override fun validateChanges(statement: Statement) {
-
-        //Ensure new tables are added
         ensureAddedBatteryControls(statement)
         ensureAddedPanDemandResponseFunctions(statement)
         ensureAddedStaticVarCompensators(statement)
         ensureAddedBatteryUnitsBatteryControls(statement)
         ensureAddedEndDevicesEndDeviceFunctions(statement)
+        ensureAddedPerLengthPhaseImpedances(statement)
+        ensureAddedPhaseImpedanceData(statement)
+        ensureRenamedPerLengthImpedance(statement)
     }
 
     override fun tearDownStatements(): List<String> =
@@ -47,7 +56,10 @@ object ChangeSet57NetworkValidator : ChangeSetValidator(DatabaseType.NETWORK_MOD
             "DELETE FROM pan_demand_response_functions;",
             "DELETE FROM static_var_compensators;",
             "DELETE FROM battery_units_battery_controls;",
-            "DELETE FROM end_devices_end_device_functions;"
+            "DELETE FROM end_devices_end_device_functions;",
+            "DELETE FROM per_length_phase_impedances;",
+            "DELETE FROM phase_impedance_data;",
+            "DELETE FROM ac_line_segments;"
         )
 
     private fun ensureAddedBatteryControls(statement: Statement) {
@@ -85,4 +97,45 @@ object ChangeSet57NetworkValidator : ChangeSetValidator(DatabaseType.NETWORK_MOD
         )
     }
 
+    private fun ensureAddedPerLengthPhaseImpedances(statement: Statement) {
+        ensureTables(statement, "per_length_phase_impedances")
+        ensureIndexes(statement, "per_length_phase_impedances_mrid")
+    }
+
+    private fun ensureAddedPhaseImpedanceData(statement: Statement) {
+        ensureTables(statement, "phase_impedance_data")
+        ensureIndexes(statement, "phase_impedance_data_per_length_phase_impedance_mrid", "phase_impedance_data_from_phase_to_phase_per_length_phase_impedance_mrid")
+    }
+
+    private fun ensureRenamedPerLengthImpedance(statement: Statement) {
+        validateRows(
+            statement, "SELECT * FROM ac_line_segments",
+            { rs ->
+                assertThat(rs.getString("mrid"), equalTo("acls1"))
+                assertThat(rs.getString("name"), equalTo(""))
+                assertThat(rs.getString("description"), equalTo(""))
+                assertThat(rs.getInt("num_diagram_objects"), equalTo(1))
+                assertThat(rs.getString("location_mrid"), equalTo("loc1"))
+                assertThat(rs.getInt("num_controls"), equalTo(10))
+                assertThat(rs.getBoolean("normally_in_service"), equalTo(true))
+                assertThat(rs.getBoolean("in_service"), equalTo(true))
+                assertThat(rs.getString("commissioned_date"), equalTo("comm1"))
+                assertThat(rs.getString("base_voltage_mrid"), equalTo("bv1"))
+                assertThat(rs.getString("per_length_impedance_mrid"), equalTo("plsi_mrid_1"))
+            },
+            { rs ->
+                assertThat(rs.getString("mrid"), equalTo("acls2"))
+                assertThat(rs.getString("name"), equalTo(""))
+                assertThat(rs.getString("description"), equalTo(""))
+                assertThat(rs.getInt("num_diagram_objects"), equalTo(1))
+                assertThat(rs.getString("location_mrid"), equalTo("loc1"))
+                assertThat(rs.getInt("num_controls"), equalTo(10))
+                assertThat(rs.getBoolean("normally_in_service"), equalTo(true))
+                assertThat(rs.getBoolean("in_service"), equalTo(true))
+                assertThat(rs.getString("commissioned_date"), equalTo("comm1"))
+                assertThat(rs.getString("base_voltage_mrid"), equalTo("bv1"))
+                assertThat(rs.getString("per_length_impedance_mrid"), equalTo("plpi_mrid_2"))
+            }
+        )
+    }
 }